refactor(packs): move complex handlers to handlers.go
- Moved final 4 handlers to handlers.go:
  - ShareMyPack (generate sharing code)
  - UnshareMyPack (remove sharing code)
  - ImportFromLighterPack (CSV import, complex handler)
  - SharedList (public shared pack endpoint)
- Added CSV-related imports to handlers.go
- packs.go now contains only error declarations (12 lines)
- All tests pass (46.1% coverage maintained)
